refactor(utils): Move analytics utilities to analytics/ subdirectory
Phase 1 (Day 2): Low-risk migration
- Moved 3 analytics files to src/utils/analytics/
- Updated 4 source imports
- Updated 2 test imports
- Tests passing (199/199 smoke tests)

Files moved:
- analytics.ts
- analytics-logger.ts
- analytics-queries.ts
