+++ src/handlers/author-search.ts
@@ -7,7 +7,7 @@ import { CacheKeyFactory } from '../services/cache-key-factory'
 import * as externalApis from '../services/external-apis'
 import { UnifiedCacheService } from '../services/unified-cache'
 import type { Env } from '../types/env'
-import { writeCacheMetrics } from '../utils/analytics'
+import { writeCacheMetrics } from '../utils/analytics/analytics'
 import { setCached } from '../utils/cache'
 
 // ============================================================================
