+++ src/utils/analytics/analytics-queries.ts
